Unit test for changeRelativePathToAbsoluteTest
