+++ .eslintrc.json
@@ -3,7 +3,8 @@
         "browser": true,
         "es6": true,
         "amd": true,
-        "node": true
+        "node": true,
+        "mocha": true
     },
     "extends": "eslint:recommended",
     "parserOptions": {
