+++ package.json
@@ -4,7 +4,9 @@
   "description": "Combine solidity files to one flat file",
   "scripts": {
     "start": "node index.js",
-    "lint": "./node_modules/.bin/eslint ."
+    "lint": "./node_modules/.bin/eslint .",
+    "test": "mocha",
+    "coveralls": "./node_modules/.bin/istanbul cover ./node_modules/mocha/bin/_mocha --report lcovonly -- -R spec && cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js && rm -rf ./coverage"
   },
   "bin": {
     "poa-solidity-flattener": "index.js"
@@ -31,6 +33,10 @@
   },
   "main": "index.js",
   "devDependencies": {
-    "eslint": "^5.5.0"
+    "coveralls": "^3.0.2",
+    "eslint": "^5.5.0",
+    "istanbul": "^0.4.5",
+    "mocha": "^5.2.0",
+    "mocha-lcov-reporter": "^1.3.0"
   }
 }
