+++ test/contracts/dep.sol
@@ -1,0 +1,6 @@
+pragma solidity ^0.4.24;
+
+
+contract Dep {
+	
+}
