+++ test/contracts/dep2.sol
@@ -1,0 +1,6 @@
+pragma solidity ^0.4.24;
+
+
+contract Dep2 {
+	
+}
