+++ test/contracts/test.sol
@@ -1,0 +1,12 @@
+pragma solidity ^0.4.24;
+
+import './dep.sol';
+import './dep2.sol';
+
+contract Test is Dep {
+	
+}
+
+contract Test2 is Dep2 {
+	
+}
