+++ test/contracts/testMock1.sol
@@ -1,0 +1,12 @@
+pragma solidity ^0.4.24;
+
+import './test/contracts/dep.sol';
+import './test/contracts/dep2.sol';
+
+contract Test is Dep {
+	
+}
+
+contract Test2 is Dep2 {
+	
+}
