+++ test/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-unused-vars
+const changeRelativePathToAbsoluteTest = require('./helpers/change-relative-path-to-absolute.js')
